requirements.txt
requests
pandas
numpy
lightgbm
scikit-learn
streamlit>=1.30.0
plotly>=5.18.0
tqdm
google-genai
